Discard source file names from the tweak file
Filter source file names from the tweak file unless they are related
to a `cms.Source("LHESource", ...)`. Only keep the first
`MaxFilesToKeep` for each attribute of interest.

Include tweaks for parameters `process.mix.input.fileNames` and `process.
mixData.input.fileNames`.

Some request configurations provide a huge list of file names that will
be overridden when the job is processed by `ReqMgr2`. Sometimes it is not possible
to upload the configuration to the configuration database because the file size
bypasses the limits. Reduce the size by removing unnecessary content.